Update README.md
It is an medical website designed with html,css and bootstrap

This project is a responsive medical website that has been developed using HTML, CSS, and the Bootstrap framework. As a result, it provides a user-friendly interface for users to access medical information, find healthcare professionals, and book appointments all from one place. With a responsive design, the site can be viewed on a variety of devices, such as laptops, tablets, and smartphones, providing an optimal viewing experience.